almost compilable

+++ util/context.go
@@ -5,6 +5,6 @@ import (
 	"time"
 )
 
-func GetTimeoutContext(second int64) (context.Context, context.CancelFunc) {
+func GetTimeoutContext(second int) (context.Context, context.CancelFunc) {
 	return context.WithTimeout(context.Background(), time.Duration(second)*time.Second)
 }
